Removed geomean files and fixhed pheatmap plotting
R package pheatmap package writes either to stdout or assigns to single pdf when
you call the function.
I want to output heatmaps that have multiple pages per pdf. Makes it less confusing
by grouping related figures together. Assigning pheatmap object to a variable will
print it stdout and save it to random Rplots.pdf and save it to the object. It also
messes up the open devices.

Changed saving pheatmap output to having pdf open device so it writes to it.